+++ field.py
@@ -1,0 +1,7 @@
+
+
+class Field:
+    '''
+    Class to switch from 3D to collapsed notation by
+    defining the __getitem__ magic method
+    '''
